Started work on a new version of the web app
The original web app grew overly complex as I added things to it. Also, the connection between the various components wasn't terribly clear.

The aim with the new app is to have a clearer separation between the SNAPS computation; the data input form; the user view; and the controller that interfaces between the other elements.